+++ remix-tests/src/fs.js
@@ -1,4 +1,5 @@
 // Extend fs
+
 var fs = require('fs')
 const path = require('path')
